% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesSUR_internal}
\alias{BayesSUR_internal}
\title{BayesSUR_internal}
\arguments{
\item{dataFile}{path to data file}

\item{outFilePath}{path to where the output is to be written}

\item{nIter}{number of iterations}

\item{nChains}{number of parallel chains to run

NOTE THAT THIS IS BASICALLY JUST A WRAPPER}
}
\description{
Run a SUR Bayesian sampler -- internal function
}
