% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpd.R
\name{elpd}
\alias{elpd}
\title{prediction accuracy}
\arguments{
\item{object}{the object from the runSUR}

\item{method}{the name of the prediction accuracy index. Default is the "\code{LOO}". The other index is the "\code{WAIC}".}
}
\description{
Measure the prediction accuracy by the elpd (expected log pointwise predictive density). One method to estimate elpd is by the The Bayesian LOO estimate of out-of-sample predictive fit. The other method the WAIC (widely applicable information criterion).
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
\references{
Vehtari, A., Gelman, A., Gabry, J. (2017). \emph{Practical Bayesian model evaluation using leave-one-out cross-validation and WAIC.} Statistics and Computing, 27(5): 1413–1432.
}
