% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_GDSC.R
\docType{data}
\name{example_GDSC}
\alias{example_GDSC}
\title{Sample dataset from the GDSC database to test the package}
\format{An object of class \code{list} of length 3.}
\usage{
example_GDSC
}
\description{
Contains 'n' observations of 's' continuous responses
and 'p' covariates
}
\details{
n = 498 ; p = 850 ; s = 7

Loading the data will load the associated 
blockList and strucutreGraph
needed to run the algorithm.
}
\keyword{datasets}
