% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimator.R
\name{getEstimator}
\alias{getEstimator}
\title{getEstimator}
\usage{
getEstimator(object, estimator = "gamma")
}
\arguments{
\item{object}{fitted "runSUR" model}

\item{estimator}{the name of one estimator. Default is the latent indicator estimator "gamma". Other options "beta" and "Gy" correspond the posterior means of coefficient matrix and response graph, respectively}
}
\description{
Extract results from the object of fitted Bayesian Seemingly Unrelated Regression
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
