% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcDiag.R
\name{mcmcDiag}
\alias{mcmcDiag}
\title{getEstimator}
\usage{
mcmcDiag(object, nbloc = 4)
}
\arguments{
\item{object}{fitted "runSUR" model}

\item{nbloc}{number of splits for the last half iterations after substracting burn-in length}
}
\description{
Diagnose the convergence of the MCMC iterations
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
