% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEstimator.R
\name{plotEstimator}
\alias{plotEstimator}
\title{plotEstimator}
\usage{
plotEstimator(object, colorScale.gamma = grey((100:0)/100))
}
\arguments{
\item{object}{fitted "runSUR" model}

\item{colorScale.gamma}{value palette for gamma}

\item{colorScale.betay}{value palette for beta}
}
\description{
Plot the estimators from the object of fitted Bayesian Seemingly Unrelated Regression
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
