% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotManhattan.R
\name{plotManhattan}
\alias{plotManhattan}
\title{Manhattan}
\usage{
plotManhattan(object, x.loc = FALSE, show.marker = NULL, xlab1 = "",
  ylab1 = "mPIP", xlab2 = "features", ylab2 = "No. of responses",
  threshold = 0.5, show.all.xlab = FALSE, las = 0, cex.axis = 1)
}
\arguments{
\item{object}{the object from the runSUR}

\item{x.loc}{a vector of features distance}

\item{show.marker}{a vector of covariate names which are shown in the Manhattan plot}

\item{xlab1}{a title for the x axis of Manhattan plot for marginal posterior inclusion probabilities}

\item{ylab1}{a title for the y axis of Manhattan plot for marginal posterior inclusion probabilities}

\item{xlab2}{a title for the x axis of Manhattan plot for numbers of responses}

\item{ylab2}{a title for the y axis of Manhattan plot for numbers of responses}

\item{threshold}{threshold for showing number of response variables significantly associated with each feature}

\item{show.all.xlab}{logical value for showing all labels on the x axis. The defaul value is to show 5 labels on the x axis}

\item{las}{graphical parameter of plot.default}

\item{cex.axis}{graphical parameter of plot.default}
}
\description{
Manhattan plot
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
