% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{plotNetwork}
\usage{
plotNetwork(object, GraphResponse = NULL, MatrixGamma = NULL,
  PmaxCovariate = 0.5, PmaxResponse = 0.5, nodesizeCovariate = 15,
  nodesizeResponse = 25, no.isolates = FALSE, lineup = 0.8)
}
\arguments{
\item{object}{fitted "runSUR" model}

\item{GraphResponse}{A matrix or dataframe of the relationship of multiple response variables. Default is "NULL" and to extrate it from object of class inheriting from "runSUR"}

\item{MatrixGamma}{A matrix or dataframe of the latent indicator variable. Default is "NULL" and to extrate it from object of class inheriting from "runSUR"}

\item{PmaxCovariate}{cutpoint for thresholding the estimated latent indicator variable. Default is 0.5}

\item{PmaxResponse}{cutpoint for thresholding the learning structure matrix of multiple response variables. Default is 0.5}

\item{nodesizeCovariate}{node size of covariates in the output graph. Default is 15}

\item{no.isolates}{remove isolated nodes from responses graph and Full graph, may get problem if there are also isolated covariates}

\item{lineup}{A ratio of the heights between responses' area and covariates'}

\item{nodesizeCovariate}{node size of response variables in the output graph. Default is 25}
}
\description{
Network representation of the associations between responses and features
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
