% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResponseGraph.R
\name{plotResponseGraph}
\alias{plotResponseGraph}
\title{plotResponseGraph}
\usage{
plotResponseGraph(object, PmaxResponse = 0.5, PtrueResponse = NULL,
  response.name = NULL)
}
\arguments{
\item{object}{fitted "runSUR" model}

\item{PmaxResponse}{cutpoint for thresholding the learning structure matrix of multiple response variables. Default is 0.5}

\item{PtrueResponse}{true adjacency matrix for the structure of multiple response variables}
}
\description{
Show the relationship between responses
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
