% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BayesSUR.R
\name{print.BayesSUR}
\alias{print.BayesSUR}
\title{print a short summary the fitted model}
\usage{
\method{print}{BayesSUR}(x, Pmax = 0.5, ...)
}
\arguments{
\item{x}{an object of class "BayesSUR".}

\item{Pmax}{cutpoint for thresholding the marginal posterior probability of inclusion (mPIP). Default is 0.5}

\item{...}{other arguments}
}
\description{
Print a short summary the fitted model generated by \code{BayesSUR()}, which is an object of class "BayesSUR".
}
\examples{
\donttest{
data("example_eQTL", package = "BayesSUR")
hyperpar = list( a_w = 2 , b_w = 5 )

fit <- BayesSUR(Y = example_eQTL[["blockList"]][[1]], 
              X = example_eQTL[["blockList"]][[2]],
              data = example_eQTL[["data"]], outFilePath = "results/",
              nIter = 1000, nChains = 2, gammaPrior = "hotspot",
              hyperpar = hyperpar, tmpFolder = "tmp/" )

## check output
# show the print information
print(fit)
}

}
