% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BayesSUR.R
\name{plot}
\alias{plot}
\alias{plot.BayesSUR}
\title{plotEstimator}
\usage{
\method{plot}{BayesSUR}(object, which = c(1L:5L))
}
\arguments{
\item{object}{fitted "runSUR" model.}

\item{which}{if a subset of the plots is required, specify a subset of the numbers 1:5 which are plots of estimators, response graph, network, manhattan and MCMC diagnosis, respectively.}
}
\description{
Six available plots from the object of fitted Bayesian Seemingly Unrelated Regression
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
