% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEstimator.R
\name{plotEstimator}
\alias{plotEstimator}
\title{plotEstimator}
\usage{
plotEstimator(object, colorScale.gamma = grey((100:0)/100),
  colorScale.beta = c("blue", "white", "red"), fig.tex = FALSE)
}
\arguments{
\item{object}{fitted "runSUR" model}

\item{colorScale.gamma}{value palette for gamma}

\item{colorScale.beta}{A vector of three colors for diverging color schemes}

\item{fig.tex}{print the figure through LaTex. Default is "FALSE"}
}
\description{
Plot the estimators from the object of fitted Bayesian Seemingly Unrelated Regression
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
