% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEstimator.R
\name{plotEstimator}
\alias{plotEstimator}
\title{plotEstimator}
\usage{
plotEstimator(object, greyscale = grey((100:0)/100))
}
\arguments{
\item{object}{fitted "runSUR" model}

\item{greyscale}{grey value palette}
}
\description{
Plot the estimators from the object of fitted Bayesian Seemingly Unrelated Regression
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
