% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMCdiag.R
\name{plotMCMCdiag}
\alias{plotMCMCdiag}
\title{getEstimator}
\usage{
plotMCMCdiag(object, nbloc = 3, ...)
}
\arguments{
\item{object}{fitted "runSUR" model}

\item{nbloc}{number of splits for the last half iterations after substracting burn-in length}

\item{...}{Other parameters in the function \code{plot.default.R} file for the penals of the log-likelihood and model size}
}
\description{
Diagnose the convergence of the MCMC iterations
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
