% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResponseGraph.R
\name{plotResponseGraph}
\alias{plotResponseGraph}
\title{plotResponseGraph}
\usage{
plotResponseGraph(object, PmaxResponse = 0.5, PtrueResponse = NULL,
  name.responses = NA, edge.weight = FALSE, label.color = "black",
  node.size = 30, node.color = "dodgerblue", ...)
}
\arguments{
\item{object}{fitted "runSUR" model}

\item{PmaxResponse}{cutpoint for thresholding the learning structure matrix of multiple response variables. Default is 0.5}

\item{PtrueResponse}{true adjacency matrix for the structure of multiple response variables}

\item{name.responses}{A vector for the node names. The default is "NA" only to show the locations. Value "auto" show the response names from the orginal data.}

\item{edge.weight}{draw weighted edges after thresholding at 0.5. The defaul value "FALSE" is not to draw weigthed edges}

\item{label.color}{label color. Default is "black"}

\item{node.size}{node size. Default is 30}

\item{node.color}{node color. Default is "dodgerblue}

\item{...}{Other parameters in the function \code{plot.default.R} file}
}
\description{
Show the relationship between responses
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
