% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BayesSUR.R
\name{summary.BayesSUR}
\alias{summary.BayesSUR}
\title{summary}
\usage{
\method{summary}{BayesSUR}(object, PmaxPredictor = 0.5)
}
\arguments{
\item{object}{fitted "runSUR" model}

\item{PmaxPredictor}{cutpoint for thresholding the estimated latent indicator variable. Default is 0.5}
}
\description{
Summarize the results from fitted model
}
\details{
BayesSUR -- Bayesian Seemingly Unrelated Regression
}
